Fix game freezing when playing song during fadeout
Apparently, SDL_mixer's Mix_FadeInMusic() will block the thread if there
is already a fadeout happening. To prevent this, if someone requests a
play() while a fadeout is already happening, we'll remember their song,
fade out quicker (unless it's from the musicfadein command), and then
play it when the fadeout is done.

We fade out quicker because it's too jarring if we stop the fadeout, and
too long if we just let the fadeout happen for its complete length.
Except for the musicfadein command, so custom levels can make their own
niceplay() if they're clever enough.